Try moving functions used by prop pack to a non-OLI util module
Use attempt_import for requests

Use attempt_import for requests

Try distributing tests to relevant module
Remove unused imports

Fix imports
