+++ watertap/core/util/periodic_table.csv
