+++ watertap/property_models/multicomp_aq_sol_prop_pack.py
@@ -78,7 +78,7 @@ from idaes.core.util.exceptions import (
 )
 import idaes.core.util.scaling as iscale
 from watertap.core.util.scaling import transform_property_constraints
-from watertap.tools.oli_api.util.watertap_to_oli_helper_functions import (
+from watertap.core.util.chemistry import (
     get_charge,
     get_molar_mass_quantity,
 )
