+++ watertap/tools/oli_api/client.py
@@ -50,11 +50,15 @@ __author__ = "Adam Atia, Adi Bannady, Paul Vecchiarelli"
 import logging
 
 import sys
-import requests
 import json
 import time
 
+from pyomo.common.dependencies import attempt_import
+
+requests, requests_available = attempt_import("requests", defer_check=False)
+
 from watertap.tools.oli_api.util.watertap_to_oli_helper_functions import get_oli_name
+
 
 _logger = logging.getLogger(__name__)
 handler = logging.StreamHandler()
