+++ watertap/tools/oli_api/util/tests/test_watertap_to_oli_helper_functions.py
@@ -17,9 +17,6 @@ from watertap.tools.oli_api.util.watertap_to_oli_helper_functions import (
     get_oli_names,
     oli_reverse_lookup,
     names_db,
-    get_charge,
-    get_molar_mass,
-    get_molar_mass_quantity,
 )
 from pyomo.environ import units as pyunits, value
 from pyomo.util.check_units import assert_units_equivalent
@@ -44,17 +41,6 @@ def test_watertap_to_oli(input_name: str, expected_output: str):
 
 
 @pytest.mark.unit
-def test_case_exception():
-    with pytest.raises(IOError) as excinfo:
-        val = "abc_2+"
-        watertap_to_oli(val)
-    assert (
-        str(excinfo.value)
-        == f" At least 1 uppercase letter is required to specify a molecule, not '{val}'."
-    )
-
-
-@pytest.mark.unit
 def test_charge_exceptions():
     with pytest.raises(IOError) as excinfo:
         val = "Na_2#"
@@ -64,68 +50,16 @@ def test_charge_exceptions():
         == "Only + and - are valid charge indicators and neither was provided in 'Na_2#'."
     )
 
-    with pytest.raises(
-        IOError,
-        match="Charge sign could not be determined from the string 'target_ion'",
-    ):
-        get_charge("target_ion")
-    with pytest.raises(
-        IOError, match="Charge could not be determined from the string 'my_target_ion'"
-    ):
-        get_charge("my_target_ion")
-
 
 @pytest.mark.unit
-def test_get_charge():
-    z = {
-        "NaCl": 0,
-        "Na_+": 1,
-        "Cl_-": -1,
-        "Ca_2+": 2,
-        "SO4_2-": -2,
-    }
-    for solute_name, charge_value in z.items():
-        assert get_charge(solute_name) == charge_value
-
-
-@pytest.mark.unit
-def test_get_mw():
-    z = {
-        "NaCl": 58.44,
-        "Na_+": 22.99,
-        "Cl_-": 35.45,
-        "Ca_2+": 40.08,
-        "SO4_2-": 96.066,
-    }
-    for solute_name, mw_value in z.items():
-        assert get_molar_mass(solute_name) == mw_value
-
-
-@pytest.mark.unit
-def test_get_mw_exception():
-    with pytest.raises(
-        IOError, match="Molecular weight data could not be found for foo."
-    ):
-        get_molar_mass("foo")
-
-
-@pytest.mark.unit
-def test_get_mw_quantity():
-    z = {
-        "NaCl": 58.44,
-        "Na_+": 22.99,
-        "Cl_-": 35.45,
-        "Ca_2+": 40.08,
-        "SO4_2-": 96.066,
-    }
-    for solute_name, mw_value in z.items():
-        result = get_molar_mass_quantity(solute_name)
-        assert value(result) == mw_value / 1000
-        assert_units_equivalent(result, pyunits.kg / pyunits.mol)
-
-        result2 = get_molar_mass_quantity(solute_name, units=pyunits.g / pyunits.mol)
-        assert value(result2) == mw_value
-        assert_units_equivalent(result2, pyunits.g / pyunits.mol)
+def test_case_exception():
+    with pytest.raises(IOError) as excinfo:
+        val = "abc_2+"
+        watertap_to_oli(val)
+    assert (
+        str(excinfo.value)
+        == f" At least 1 uppercase letter is required to specify a molecule, not '{val}'."
+    )
 
 
 @pytest.mark.unit
